Defer scheduling to twisted
It was not possible to run the twisted reactor multiple times.
Usual solution is to run it in multiprocessing, but that kept crashing on Mac
(because of wxPython and pyobj, after experimentation)

This means that there is now two kinds of scheduling, and apscheduler seems redundant.
May be worth to use a wx timer for the mac theme detection?